requirements.txt
# do not add versions randomly

python-telegram-bot[webhooks]==20.7
openai
python-dotenv==1.0.1
asgiref==3.7.2
Django==5.0.2
psycopg2-binary==2.9.9
dj-database-url
Pillow
gunicorn
whitenoise==6.6.0
pydantic
stripe
django-json-widget
markdown2
whitenoise
watchdog
black
pyyaml
attrs
django-extensions
cryptography
pre-commit
sentry-sdk[django]
reportlab
pdf2image
Pillow
boto3
django-storages
requests
pynacl
django-cors-headers